requirements.txt
requests>=2.32.0
PyYAML>=6.0.0
pyshark>=0.6.0
Flask>=3.0.0
